ace-spider/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
python-dotenv>=1.0.0
openai>=1.0.0
apscheduler>=3.10.0
openpyxl>=3.1.0
